requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jsoncodable==0.1.7
jsonpickle==3.3.0
k-selenium-cookies==0.0.4
kproxy==0.0.1
noraise==0.0.16
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.3
requests==2.32.3
requests-file==2.1.0
selenium==4.25.0
selenium-browser==0.0.15
selenium-firefox==2.0.8
setuptools==75.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
xpath-utils==0.0.3
